improve motor speed (+ latency test)
Regarding speed: the motor seems to have some smarts, and only starts going full speed once it senses that a human is turning the knob and the motor is near the target position. (n%2)  is added to position as a noise source, that seems to trick the motor to engage the full speed; clean back-and-forth didn't work even with static position being sent at a high rate.

Latency test:: change the color of led before sending a command with a fresh position.